refactor(application-header): extract launchpad category component
Move category rendering into a focused component and apply the extra spacing to subsequent category hosts.
