+++ projects/element-ng/application-header/launchpad/si-launchpad-factory.component.ts
@@ -4,20 +4,14 @@
  */
 import { A11yModule } from '@angular/cdk/a11y';
 import { booleanAttribute, Component, computed, inject, input, output } from '@angular/core';
-import { ActivatedRoute, RouterLink, RouterLinkActive } from '@angular/router';
 import { elementCancel, elementDown2 } from '@siemens/element-icons';
 import { addIcons, SiIconComponent } from '@siemens/element-ng/icon';
 import { SiLinkModule } from '@siemens/element-ng/link';
 import { SiTranslatePipe, t, TranslatableString } from '@siemens/element-translate-ng/translate';
 
 import { SiApplicationHeaderComponent } from '../si-application-header.component';
-import { SiLaunchpadAppComponent } from './si-launchpad-app.component';
-import { App, AppCategory } from './si-launchpad.model';
-
-export interface FavoriteChangeEvent {
-  app: App;
-  favorite: boolean;
-}
+import { SiLaunchpadCategoryComponent } from './si-launchpad-category.component';
+import { App, AppCategory, FavoriteChangeEvent } from './si-launchpad.model';
 
 @Component({
   selector: 'si-launchpad-factory',
@@ -25,10 +19,8 @@ export interface FavoriteChangeEvent {
     A11yModule,
     SiLinkModule,
     SiTranslatePipe,
-    SiLaunchpadAppComponent,
-    SiIconComponent,
-    RouterLinkActive,
-    RouterLink
+    SiLaunchpadCategoryComponent,
+    SiIconComponent
   ],
   templateUrl: './si-launchpad-factory.component.html',
   styleUrl: './si-launchpad-factory.component.scss'
@@ -128,15 +120,14 @@ export class SiLaunchpadFactoryComponent {
   );
   protected readonly hasFavorites = computed(() => this.favorites().length > 0);
   protected readonly icons = addIcons({ elementDown2, elementCancel });
-  protected readonly activatedRoute = inject(ActivatedRoute, { optional: true });
   private header = inject(SiApplicationHeaderComponent);
 
   protected closeLaunchpad(): void {
     this.header.closeLaunchpad();
   }
 
-  protected toggleFavorite(app: App, favorite: boolean): void {
-    this.favoriteChange.emit({ app, favorite });
+  protected toggleFavorite(event: FavoriteChangeEvent): void {
+    this.favoriteChange.emit(event);
   }
 
   protected escape(): void {
@@ -145,13 +136,5 @@ export class SiLaunchpadFactoryComponent {
 
   protected isCategories(items: App[] | AppCategory[]): items is AppCategory[] {
     return items.some(item => 'apps' in item);
-  }
-
-  protected isFavoriteToggleDisabled(app: App): boolean {
-    if ('_noFavorite' in app) {
-      return !!app._noFavorite;
-    } else {
-      return false;
-    }
   }
 }
